backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
celery>=5.4.0
redis>=5.0.5
google-generativeai>=0.5.2
python-dotenv>=1.0.1
requests>=2.32.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
